+++ src/algorithms/local_search/local_search.h
@@ -29,6 +29,7 @@ template <class Route,
           class IntraMixedExchange,
           class IntraRelocate,
           class IntraOrOpt,
+          class IntraTwoOpt,
           class PDShift,
           class RouteExchange>
 class LocalSearch {